modift:[User] jwt 문제 해결

+++ src/main/java/com/example/demo/src/user/model/PostLoginRes.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class PostLoginRes {
     private String jwt;
-    private int expiration;
+    private long expiration;
     private String refresh;
     private int userIdx;
 }
